Ver: 5.4b
Finished new algorithm of backups:
Now we copy files to the bucket and later we remove the old ones. 

Also, added a secondary script which helps regenerate VBM files in case they are corrupted or broken since in every upload we'll mash the .vbm since they have no date in the file name.